+++ CHANGELOG.md
@@ -7,8 +7,15 @@ This release supports API1200 minimally where we can use OneView API1200 with th
 
 ### Oneview Features supported
 - Enclosure
+- Enclosure Group
+- FC Network
+- FCoE Network
+- Interconnect Type
+- Logical Enclosure
+- Logical Interconnect Group
 - Network Set
 - Server Profile
+- Server Profile Template
 - Storage Pool
 - Storage System
 - Storage Volume
@@ -40,6 +47,8 @@ This version supports OneView appliances with version 4.2 using the OneView REST
 - Scope
 - Server hardware
 - Server hardware type
+- Server profile
+- Server profile template
 - Storage pool
 - Storage system
 - Storage volume